remove unused and refactor components
