+++ ui/app/dashboard/_components/dashboard-client.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import {
-	AlertTriangle,
 	BarChart3,
 	Clock,
 	Database,
@@ -15,47 +14,18 @@ import {
 	Upload,
 } from "lucide-react";
 import { useEffect, useRef, useState } from "react";
-import {
-	CartesianGrid,
-	Legend,
-	Line,
-	LineChart,
-	ResponsiveContainer,
-	Tooltip,
-	XAxis,
-	YAxis,
-} from "recharts";
 import { Badge } from "@/components/ui/badge";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import {
-	Table,
-	TableBody,
-	TableCell,
-	TableHead,
-	TableHeader,
-	TableRow,
-} from "@/components/ui/table";
-import {
-	type BrokerStatus,
-	formatBytes,
-	formatCount,
-	formatUptime,
-	type NodeInfo,
-} from "@/lib/api";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent } from "@/components/ui/card";
+import { formatBytes, formatCount, formatUptime } from "@/lib/format";
 import { getBrokerOverview } from "@/lib/services/broker";
+import type { BrokerStatus, NodeInfo } from "@/lib/types";
+import { ClusterNodesTable } from "./cluster-nodes-table";
+import { type ChartPoint, MetricCharts } from "./metric-charts";
 
 const POLL_MS = 5_000;
 const POLL_S = POLL_MS / 1000;
 const MAX_POINTS = 40;
-
-interface ChartPoint {
-	time: string;
-	sessions: number;
-	msgsIn: number;
-	msgsOut: number;
-	bytesIn: number;
-	bytesOut: number;
-}
 
 function now(): string {
 	return new Date().toLocaleTimeString([], {
@@ -88,6 +58,41 @@ function nodeToStatus(node: NodeInfo): BrokerStatus {
 		protocol_errors: 0,
 		packet_errors: 0,
 	};
+}
+
+interface NodePillProps {
+	label: string;
+	active: boolean;
+	isLeader?: boolean;
+	onClick: () => void;
+}
+
+function NodePill({ label, active, isLeader, onClick }: NodePillProps) {
+	return (
+		<button
+			type="button"
+			onClick={onClick}
+			className={`flex items-center gap-1.5 text-xs px-3 py-2 min-h-10 rounded-full border transition-colors whitespace-nowrap ${
+				active
+					? "bg-flux-blue text-white border-flux-blue"
+					: "bg-flux-card text-flux-text-muted border-flux-card-border hover:bg-flux-hover hover:text-flux-text"
+			}`}
+		>
+			{isLeader !== undefined && (
+				<span
+					className={`inline-block w-1.5 h-1.5 rounded-full ${active ? "bg-white/70" : "bg-flux-green"}`}
+				/>
+			)}
+			{label}
+			{isLeader && (
+				<span
+					className={`text-[10px] ${active ? "text-white/70" : "text-flux-blue"}`}
+				>
+					★
+				</span>
+			)}
+		</button>
+	);
 }
 
 export default function DashboardClient() {
@@ -123,6 +128,7 @@ export default function DashboardClient() {
 				const ts = now();
 				const histories = historiesRef.current;
 				const prev = prevRef.current;
+
 				function makePoint(
 					sessions: number,
 					msgsRx: number,
@@ -202,20 +208,6 @@ export default function DashboardClient() {
 	const historyKey = selectedNodeId ?? "";
 	const displayHistory = historiesRef.current[historyKey] ?? [];
 
-	const gridColor = "var(--flux-grid)";
-	const axisColor = "var(--flux-text-muted)";
-	const tooltipBg = "var(--flux-card)";
-	const tooltipBdr = "var(--flux-card-border)";
-	const tooltipTxt = "var(--flux-text)";
-
-	const tooltipStyle = {
-		backgroundColor: tooltipBg,
-		border: `1px solid ${tooltipBdr}`,
-		borderRadius: "8px",
-		color: tooltipTxt,
-		fontSize: 12,
-	};
-
 	const statCards = displayStatus
 		? [
 				{ label: "Active Connections", value: formatCount(displayStatus.sessions), icon: BarChart3 },
@@ -291,6 +283,7 @@ export default function DashboardClient() {
 					</Badge>
 				</div>
 			</div>
+
 			{error && (
 				<div
 					role="alert"
@@ -299,6 +292,7 @@ export default function DashboardClient() {
 					⚠ {error}
 				</div>
 			)}
+
 			<div className="flex items-center gap-2 overflow-x-auto pb-1">
 				<Share2 className="w-4 h-4 text-flux-text-muted shrink-0" />
 				<NodePill
@@ -316,6 +310,7 @@ export default function DashboardClient() {
 					/>
 				))}
 			</div>
+
 			<div className="grid grid-cols-2 sm:grid-cols-4 gap-4">
 				{statCards.map((card) => {
 					const Icon = card.icon;
@@ -341,372 +336,24 @@ export default function DashboardClient() {
 					);
 				})}
 			</div>
-			<div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-				<Card className="border-flux-card-border bg-flux-card">
-					<CardHeader className="pb-2">
-						<CardTitle className="text-base text-flux-text">
-							Message Traffic Trends
-						</CardTitle>
-						<p className="text-xs text-flux-text-muted">
-							{scopeLabel} · msgs/s in &amp; out · polled every {POLL_S}s
-						</p>
-					</CardHeader>
-					<CardContent>
-						<div
-							className="relative"
-							role="img"
-							aria-label={`Message traffic trend chart for ${scopeLabel}`}
-						>
-							{!loaded && <ChartLoadingSkeleton height={260} />}
-							<ResponsiveContainer width="100%" height={260}>
-								<LineChart
-									data={displayHistory}
-									margin={{ top: 4, right: 8, left: 0, bottom: 0 }}
-								>
-									<CartesianGrid strokeDasharray="3 3" stroke={gridColor} />
-									<XAxis
-										dataKey="time"
-										stroke={axisColor}
-										tick={{ fontSize: 10 }}
-										interval="preserveStartEnd"
-									/>
-									<YAxis
-										stroke={axisColor}
-										tick={{ fontSize: 10 }}
-										allowDecimals={false}
-										unit="/s"
-									/>
-									<Tooltip
-										contentStyle={tooltipStyle}
-										formatter={(v) => [
-											typeof v === "number" ? `${v.toFixed(1)}/s` : "",
-										]}
-									/>
-									<Legend wrapperStyle={{ fontSize: 11 }} />
-									<Line
-										type="monotone"
-										dataKey="msgsIn"
-										name="In"
-										stroke="var(--flux-green)"
-										strokeWidth={2}
-										dot={false}
-										isAnimationActive={false}
-									/>
-									<Line
-										type="monotone"
-										dataKey="msgsOut"
-										name="Out"
-										stroke="var(--flux-blue)"
-										strokeWidth={2}
-										dot={false}
-										isAnimationActive={false}
-									/>
-								</LineChart>
-							</ResponsiveContainer>
-							<p className="sr-only">{trafficSummary}</p>
-							{error && <ChartErrorOverlay />}
-						</div>
-					</CardContent>
-				</Card>
 
-				<Card className="border-flux-card-border bg-flux-card">
-					<CardHeader className="pb-2">
-						<CardTitle className="text-base text-flux-text">
-							Bandwidth
-						</CardTitle>
-						<p className="text-xs text-flux-text-muted">
-							{scopeLabel} · bytes/s in &amp; out
-						</p>
-					</CardHeader>
-					<CardContent>
-						<div
-							className="relative"
-							role="img"
-							aria-label={`Bandwidth trend chart for ${scopeLabel}`}
-						>
-							{!loaded && <ChartLoadingSkeleton height={260} />}
-							<ResponsiveContainer width="100%" height={260}>
-								<LineChart
-									data={displayHistory}
-									margin={{ top: 4, right: 8, left: 0, bottom: 0 }}
-								>
-									<CartesianGrid strokeDasharray="3 3" stroke={gridColor} />
-									<XAxis
-										dataKey="time"
-										stroke={axisColor}
-										tick={{ fontSize: 10 }}
-										interval="preserveStartEnd"
-									/>
-									<YAxis
-										stroke={axisColor}
-										tick={{ fontSize: 10 }}
-										allowDecimals={false}
-										tickFormatter={(v: number) => formatBytes(v)}
-									/>
-									<Tooltip
-										contentStyle={tooltipStyle}
-										formatter={(v) => [
-											typeof v === "number" ? `${formatBytes(v)}/s` : "",
-										]}
-									/>
-									<Legend wrapperStyle={{ fontSize: 11 }} />
-									<Line
-										type="monotone"
-										dataKey="bytesIn"
-										name="In"
-										stroke="var(--flux-orange)"
-										strokeWidth={2}
-										dot={false}
-										isAnimationActive={false}
-									/>
-									<Line
-										type="monotone"
-										dataKey="bytesOut"
-										name="Out"
-										stroke="var(--flux-purple)"
-										strokeWidth={2}
-										dot={false}
-										isAnimationActive={false}
-									/>
-								</LineChart>
-							</ResponsiveContainer>
-							<p className="sr-only">{bandwidthSummary}</p>
-							{error && <ChartErrorOverlay />}
-						</div>
-					</CardContent>
-				</Card>
+			<MetricCharts
+				displayHistory={displayHistory}
+				loaded={loaded}
+				error={error}
+				scopeLabel={scopeLabel}
+				selectedNodeId={selectedNodeId}
+				pollSeconds={POLL_S}
+				trafficSummary={trafficSummary}
+				bandwidthSummary={bandwidthSummary}
+				sessionsSummary={sessionsSummary}
+			/>
 
-				{selectedNodeId === null && (
-					<Card className="border-flux-card-border bg-flux-card lg:col-span-2">
-						<CardHeader className="pb-2">
-							<CardTitle className="text-base text-flux-text">
-								Active Connections
-							</CardTitle>
-							<p className="text-xs text-flux-text-muted">
-								Cluster · connected clients over time
-							</p>
-						</CardHeader>
-						<CardContent>
-							<div
-								className="relative"
-								role="img"
-								aria-label="Active connections trend chart for cluster"
-							>
-								{!loaded && <ChartLoadingSkeleton height={200} />}
-								<ResponsiveContainer width="100%" height={200}>
-									<LineChart
-										data={displayHistory}
-										margin={{ top: 4, right: 8, left: 0, bottom: 0 }}
-									>
-										<CartesianGrid strokeDasharray="3 3" stroke={gridColor} />
-										<XAxis
-											dataKey="time"
-											stroke={axisColor}
-											tick={{ fontSize: 10 }}
-											interval="preserveStartEnd"
-										/>
-										<YAxis
-											stroke={axisColor}
-											tick={{ fontSize: 10 }}
-											allowDecimals={false}
-										/>
-										<Tooltip contentStyle={tooltipStyle} />
-										<Line
-											type="monotone"
-											dataKey="sessions"
-											name="Connections"
-											stroke="var(--flux-blue)"
-											strokeWidth={2}
-											dot={false}
-											isAnimationActive={false}
-										/>
-									</LineChart>
-								</ResponsiveContainer>
-								<p className="sr-only">{sessionsSummary}</p>
-								{error && <ChartErrorOverlay />}
-							</div>
-						</CardContent>
-					</Card>
-				)}
-			</div>
-
-			{nodes.length > 0 && (
-				<Card className="border-flux-card-border bg-flux-card">
-					<CardHeader className="pb-3">
-						<div className="flex items-center justify-between">
-							<div>
-								<CardTitle className="text-base text-flux-text">
-									Cluster Nodes
-								</CardTitle>
-								<p className="text-xs text-flux-text-muted mt-0.5">
-									{nodes.length} node{nodes.length !== 1 ? "s" : ""}
-								</p>
-							</div>
-						</div>
-					</CardHeader>
-					<CardContent className="p-0">
-						<div className="overflow-x-auto">
-							<Table>
-								<TableHeader>
-									<TableRow className="border-flux-card-border hover:bg-transparent">
-										<TableHead className="pl-6">Node</TableHead>
-										<TableHead>Address</TableHead>
-										<TableHead className="text-right">Sessions</TableHead>
-										<TableHead className="text-right">Subscriptions</TableHead>
-										<TableHead className="text-right">Msgs In</TableHead>
-										<TableHead className="text-right">Msgs Out</TableHead>
-										<TableHead className="text-right">Bytes In</TableHead>
-										<TableHead className="text-right pr-6">Uptime</TableHead>
-									</TableRow>
-								</TableHeader>
-								<TableBody>
-									{nodes.map((node) => {
-										const isSelected = selectedNodeId === node.node_id;
-										return (
-											<TableRow
-												key={node.node_id}
-												className={`border-flux-card-border transition-colors ${
-													isSelected
-														? "bg-flux-blue/10 hover:bg-flux-blue/15"
-														: "hover:bg-flux-hover"
-												}`}
-											>
-												<TableCell className="pl-6 py-4">
-													<button
-														type="button"
-														aria-pressed={isSelected}
-														aria-label={`Inspect node ${node.node_id}`}
-														onClick={() =>
-															setSelectedNodeId(
-																isSelected ? null : node.node_id,
-															)
-														}
-														className="flex w-full items-center gap-2 rounded text-left focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-flux-blue"
-													>
-														<span className="inline-block w-2 h-2 rounded-full bg-flux-green shrink-0" />
-														<span className="text-flux-text font-medium text-sm">
-															{node.node_id}
-														</span>
-														{node.is_leader && (
-															<Badge
-																variant="outline"
-																className="text-xs bg-flux-blue/10 text-flux-blue border-flux-blue/20"
-															>
-																Leader
-															</Badge>
-														)}
-													</button>
-												</TableCell>
-												<TableCell className="text-flux-text-muted font-mono text-xs py-4">
-													{node.addr}
-												</TableCell>
-												<TableCell className="text-flux-text text-sm text-right py-4">
-													{node.sessions !== undefined ? (
-														formatCount(node.sessions)
-													) : (
-														<span className="text-flux-text-muted">—</span>
-													)}
-												</TableCell>
-												<TableCell className="text-flux-text text-sm text-right py-4">
-													{node.subscriptions !== undefined ? (
-														formatCount(node.subscriptions)
-													) : (
-														<span className="text-flux-text-muted">—</span>
-													)}
-												</TableCell>
-												<TableCell className="text-flux-text text-sm text-right py-4">
-													{node.messages_received !== undefined ? (
-														formatCount(node.messages_received)
-													) : (
-														<span className="text-flux-text-muted">—</span>
-													)}
-												</TableCell>
-												<TableCell className="text-flux-text text-sm text-right py-4">
-													{node.messages_sent !== undefined ? (
-														formatCount(node.messages_sent)
-													) : (
-														<span className="text-flux-text-muted">—</span>
-													)}
-												</TableCell>
-												<TableCell className="text-flux-text text-sm text-right py-4">
-													{node.bytes_received !== undefined ? (
-														formatBytes(node.bytes_received)
-													) : (
-														<span className="text-flux-text-muted">—</span>
-													)}
-												</TableCell>
-												<TableCell className="text-flux-text-muted text-sm text-right pr-6 py-4">
-													{formatUptime(node.uptime_seconds)}
-												</TableCell>
-											</TableRow>
-										);
-									})}
-								</TableBody>
-							</Table>
-						</div>
-					</CardContent>
-				</Card>
-			)}
+			<ClusterNodesTable
+				nodes={nodes}
+				selectedNodeId={selectedNodeId}
+				onSelectNode={setSelectedNodeId}
+			/>
 		</div>
-	);
-}
-
-interface NodePillProps {
-	label: string;
-	active: boolean;
-	isLeader?: boolean;
-	onClick: () => void;
-}
-
-function ChartLoadingSkeleton({ height }: { height: number }) {
-	return (
-		<div
-			className="absolute inset-0 z-10 flex items-center justify-center rounded-md bg-flux-card"
-			style={{ height }}
-		>
-			<div className="flex flex-col items-center gap-2">
-				<div className="h-1.5 w-24 animate-pulse rounded-full bg-flux-card-border" />
-				<div className="h-1.5 w-16 animate-pulse rounded-full bg-flux-card-border" />
-			</div>
-		</div>
-	);
-}
-
-function ChartErrorOverlay() {
-	return (
-		<div className="absolute inset-0 flex flex-col items-center justify-center gap-2 rounded-md backdrop-blur-[2px] bg-flux-bg/60">
-			<AlertTriangle className="w-5 h-5 text-flux-red" />
-			<p className="text-xs font-medium text-flux-text-muted">
-				Data unavailable
-			</p>
-		</div>
-	);
-}
-
-function NodePill({ label, active, isLeader, onClick }: NodePillProps) {
-	return (
-		<button
-			type="button"
-			onClick={onClick}
-			className={`flex items-center gap-1.5 text-xs px-3 py-2 min-h-10 rounded-full border transition-colors whitespace-nowrap ${
-				active
-					? "bg-flux-blue text-white border-flux-blue"
-					: "bg-flux-card text-flux-text-muted border-flux-card-border hover:bg-flux-hover hover:text-flux-text"
-			}`}
-		>
-			{isLeader !== undefined && (
-				<span
-					className={`inline-block w-1.5 h-1.5 rounded-full ${active ? "bg-white/70" : "bg-flux-green"}`}
-				/>
-			)}
-			{label}
-			{isLeader && (
-				<span
-					className={`text-[10px] ${active ? "text-white/70" : "text-flux-blue"}`}
-				>
-					★
-				</span>
-			)}
-		</button>
 	);
 }
