+++ ui/app/layout.tsx
@@ -1,6 +1,9 @@
 import type { Metadata } from "next";
+import { Geist } from "next/font/google";
 import { Providers } from "@/components/providers";
 import "./globals.css";
+
+const geist = Geist({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
 	title: "FluxMQ Dashboard",
@@ -14,19 +17,7 @@ export default function RootLayout({
 }) {
 	return (
 		<html lang="en" suppressHydrationWarning>
-			<head>
-				<link rel="preconnect" href="https://fonts.googleapis.com" />
-				<link
-					rel="preconnect"
-					href="https://fonts.gstatic.com"
-					crossOrigin="anonymous"
-				/>
-				<link
-					href="https://fonts.googleapis.com/css2?family=Geist:wght@100..900&display=swap"
-					rel="stylesheet"
-				/>
-			</head>
-			<body className="font-geist">
+			<body className={geist.className}>
 				<Providers>{children}</Providers>
 			</body>
 		</html>
