+++ ui/lib/api.ts
@@ -1,493 +1,14 @@
-export interface BrokerStatus {
-	node_id: string;
-	is_leader: boolean;
-	cluster_mode: boolean;
-	node_count?: number;
-	sessions: number;
-	sessions_total: number;
-	connections_total: number;
-	connections_disconnections: number;
-	messages_received: number;
-	messages_sent: number;
-	publish_received: number;
-	publish_sent: number;
-	bytes_received: number;
-	bytes_sent: number;
-	subscriptions: number;
-	retained_messages: number;
-	uptime_seconds: number;
-	auth_errors: number;
-	authz_errors: number;
-	protocol_errors: number;
-	packet_errors: number;
-}
-
-export interface NodeInfo {
-	node_id: string;
-	is_leader: boolean;
-	healthy?: boolean;
-	addr: string;
-	uptime_seconds: number;
-	sessions?: number;
-	subscriptions?: number;
-	messages_received?: number;
-	messages_sent?: number;
-	bytes_received?: number;
-	bytes_sent?: number;
-}
-
-export interface SessionSubscription {
-	filter: string;
-	qos: number;
-	no_local: boolean;
-	retain_as_published: boolean;
-	retain_handling: number;
-	consumer_group?: string;
-	subscription_id?: number;
-}
-
-export interface SessionInfo {
-	client_id: string;
-	connection_name?: string;
-	node_id?: string;
-	state: string; // "connected" | "disconnected"
-	connected: boolean;
-	protocol: string;
-	version: number;
-	clean_start: boolean;
-	expiry_interval: number;
-	connected_at?: string;
-	disconnected_at?: string;
-	receive_maximum: number;
-	max_packet_size: number;
-	topic_alias_max: number;
-	request_response: boolean;
-	request_problem: boolean;
-	has_will: boolean;
-	subscription_count: number;
-	inflight_count: number;
-	offline_queue_depth: number;
-	subscriptions?: SessionSubscription[];
-}
-
-export const MOCK_SESSIONS: SessionInfo[] = [
-	{
-		client_id: "mqtt-client-001",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T07:00:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 10,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 5,
-		inflight_count: 0,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "mqtt-client-002",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt3.1.1",
-		version: 4,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T08:30:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 3,
-		inflight_count: 2,
-		offline_queue_depth: 2,
-	},
-	{
-		client_id: "mqtt-client-003",
-		state: "disconnected",
-		connected: false,
-		protocol: "mqtt3.1.1",
-		version: 4,
-		clean_start: false,
-		expiry_interval: 300,
-		connected_at: "2026-03-18T06:00:00Z",
-		disconnected_at: "2026-03-18T10:14:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 0,
-		inflight_count: 0,
-		offline_queue_depth: 7,
-	},
-	{
-		client_id: "sensor-device-001",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T09:00:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 1048576,
-		topic_alias_max: 5,
-		request_response: false,
-		request_problem: false,
-		has_will: true,
-		subscription_count: 1,
-		inflight_count: 0,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "sensor-device-002",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T09:01:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 1048576,
-		topic_alias_max: 5,
-		request_response: false,
-		request_problem: false,
-		has_will: true,
-		subscription_count: 1,
-		inflight_count: 0,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "gateway-node-alpha",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: false,
-		expiry_interval: 3600,
-		connected_at: "2026-03-17T22:00:00Z",
-		receive_maximum: 1000,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: true,
-		request_problem: true,
-		has_will: true,
-		subscription_count: 8,
-		inflight_count: 3,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "gateway-node-beta",
-		state: "disconnected",
-		connected: false,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: false,
-		expiry_interval: 3600,
-		connected_at: "2026-03-17T22:05:00Z",
-		disconnected_at: "2026-03-18T09:52:00Z",
-		receive_maximum: 1000,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: true,
-		request_problem: true,
-		has_will: true,
-		subscription_count: 0,
-		inflight_count: 0,
-		offline_queue_depth: 14,
-	},
-	{
-		client_id: "logger-service",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt3.1.1",
-		version: 4,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T00:00:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 12,
-		inflight_count: 0,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "edge-reporter-01",
-		state: "disconnected",
-		connected: false,
-		protocol: "mqtt3.1.1",
-		version: 4,
-		clean_start: false,
-		expiry_interval: 300,
-		connected_at: "2026-03-18T05:00:00Z",
-		disconnected_at: "2026-03-18T08:30:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 0,
-		inflight_count: 0,
-		offline_queue_depth: 3,
-	},
-	{
-		client_id: "dashboard-ingest",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T06:30:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 10,
-		request_response: true,
-		request_problem: true,
-		has_will: false,
-		subscription_count: 6,
-		inflight_count: 1,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "monitoring-agent",
-		state: "connected",
-		connected: true,
-		protocol: "mqtt3.1",
-		version: 3,
-		clean_start: true,
-		expiry_interval: 0,
-		connected_at: "2026-03-18T03:00:00Z",
-		receive_maximum: 65535,
-		max_packet_size: 0,
-		topic_alias_max: 0,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 4,
-		inflight_count: 0,
-		offline_queue_depth: 0,
-	},
-	{
-		client_id: "config-pusher",
-		state: "disconnected",
-		connected: false,
-		protocol: "mqtt5",
-		version: 5,
-		clean_start: false,
-		expiry_interval: 86400,
-		connected_at: "2026-03-17T18:00:00Z",
-		disconnected_at: "2026-03-18T01:00:00Z",
-		receive_maximum: 100,
-		max_packet_size: 65536,
-		topic_alias_max: 0,
-		request_response: false,
-		request_problem: false,
-		has_will: false,
-		subscription_count: 2,
-		inflight_count: 0,
-		offline_queue_depth: 0,
-	},
-];
-
-const NODE_RATES: Record<
-	string,
-	{
-		msgsIn: number;
-		msgsOut: number;
-		bytesIn: number;
-		bytesOut: number;
-		sessions: number;
-	}
-> = {
-	"broker-1": {
-		msgsIn: 12,
-		msgsOut: 11,
-		bytesIn: 4_800,
-		bytesOut: 4_200,
-		sessions: 20,
-	},
-	"broker-2": {
-		msgsIn: 7,
-		msgsOut: 6,
-		bytesIn: 2_800,
-		bytesOut: 2_400,
-		sessions: 12,
-	},
-	"broker-3": {
-		msgsIn: 5,
-		msgsOut: 4,
-		bytesIn: 2_000,
-		bytesOut: 1_700,
-		sessions: 10,
-	},
-};
-
-const _nodes: NodeInfo[] = [
-	{
-		node_id: "broker-1",
-		is_leader: true,
-		addr: "10.0.0.1:4001",
-		sessions: 20,
-		subscriptions: 87,
-		messages_received: 58_200,
-		messages_sent: 56_100,
-		bytes_received: 23_068_672,
-		bytes_sent: 21_233_664,
-		uptime_seconds: 86_400,
-	},
-	{
-		node_id: "broker-2",
-		is_leader: false,
-		addr: "10.0.0.2:4001",
-		sessions: 12,
-		subscriptions: 62,
-		messages_received: 41_750,
-		messages_sent: 40_320,
-		bytes_received: 17_825_792,
-		bytes_sent: 16_252_928,
-		uptime_seconds: 82_800,
-	},
-	{
-		node_id: "broker-3",
-		is_leader: false,
-		addr: "10.0.0.3:4001",
-		sessions: 10,
-		subscriptions: 38,
-		messages_received: 28_500,
-		messages_sent: 27_900,
-		bytes_received: 11_534_336,
-		bytes_sent: 10_747_904,
-		uptime_seconds: 79_200,
-	},
-];
-
-const POLL_S = 5;
-
-function aggregateStatus(): BrokerStatus {
-	return {
-		node_id: "broker-1",
-		is_leader: true,
-		cluster_mode: true,
-		node_count: _nodes.length,
-		sessions: _nodes.reduce((s, n) => s + (n.sessions ?? 0), 0),
-		subscriptions: _nodes.reduce((s, n) => s + (n.subscriptions ?? 0), 0),
-		messages_received: _nodes.reduce(
-			(s, n) => s + (n.messages_received ?? 0),
-			0,
-		),
-		messages_sent: _nodes.reduce((s, n) => s + (n.messages_sent ?? 0), 0),
-		bytes_received: _nodes.reduce((s, n) => s + (n.bytes_received ?? 0), 0),
-		bytes_sent: _nodes.reduce((s, n) => s + (n.bytes_sent ?? 0), 0),
-		sessions_total: _nodes.reduce((s, n) => s + (n.sessions ?? 0), 0) + 12,
-		connections_total: 1042,
-		connections_disconnections: 87,
-		publish_received: _nodes.reduce(
-			(s, n) => s + (n.messages_received ?? 0),
-			0,
-		),
-		publish_sent: _nodes.reduce((s, n) => s + (n.messages_sent ?? 0), 0),
-		retained_messages: 34,
-		uptime_seconds: _nodes[0].uptime_seconds,
-		auth_errors: 3,
-		authz_errors: 1,
-		protocol_errors: 2,
-		packet_errors: 0,
-	};
-}
-
-export const MOCK_STATUS: BrokerStatus = aggregateStatus();
-export const MOCK_NODES: NodeInfo[] = _nodes.map((n) => ({ ...n }));
-
-export interface MockHistoryPoint {
-	time: string;
-	sessions: number;
-	msgsIn: number;
-	msgsOut: number;
-	bytesIn: number;
-	bytesOut: number;
-}
-
-export function generateMockHistory(
-	nodeId: string | null,
-	points = 30,
-): MockHistoryPoint[] {
-	const rates = nodeId
-		? (NODE_RATES[nodeId] ?? NODE_RATES["broker-1"])
-		: {
-				msgsIn: 24,
-				msgsOut: 21,
-				bytesIn: 9_600,
-				bytesOut: 8_300,
-				sessions: 42,
-			};
-
-	const baseSessions = rates.sessions;
-	const history: MockHistoryPoint[] = [];
-	const nowMs = Date.now();
-
-	for (let i = points; i >= 1; i--) {
-		const t = new Date(nowMs - i * POLL_S * 1000);
-		const timeStr = t.toLocaleTimeString([], {
-			hour: "2-digit",
-			minute: "2-digit",
-			second: "2-digit",
-		});
-
-		const wave = Math.sin(((points - i) / points) * Math.PI) * 0.3 + 0.85;
-		history.push({
-			time: timeStr,
-			sessions: Math.round(baseSessions * (1 + (Math.random() * 0.1 - 0.05))),
-			msgsIn: Math.round(
-				rates.msgsIn * wave * (1 + (Math.random() * 0.3 - 0.15)),
-			),
-			msgsOut: Math.round(
-				rates.msgsOut * wave * (1 + (Math.random() * 0.3 - 0.15)),
-			),
-			bytesIn: Math.round(
-				rates.bytesIn * wave * (1 + (Math.random() * 0.3 - 0.15)),
-			),
-			bytesOut: Math.round(
-				rates.bytesOut * wave * (1 + (Math.random() * 0.3 - 0.15)),
-			),
-		});
-	}
-	return history;
-}
-
-export function formatCount(n: number): string {
-	if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-	if (n >= 1_000) return `${(n / 1_000).toFixed(1)}K`;
-	return String(n);
-}
-
-export function formatBytes(bytes: number): string {
-	if (bytes >= 1_073_741_824) return `${(bytes / 1_073_741_824).toFixed(1)} GB`;
-	if (bytes >= 1_048_576) return `${(bytes / 1_048_576).toFixed(1)} MB`;
-	if (bytes >= 1_024) return `${(bytes / 1_024).toFixed(1)} KB`;
-	return `${bytes} B`;
-}
-
-export function formatUptime(seconds: number): string {
-	const d = Math.floor(seconds / 86400);
-	const h = Math.floor((seconds % 86400) / 3600);
-	const m = Math.floor((seconds % 3600) / 60);
-	if (d > 0) return `${d}d ${h}h`;
-	if (h > 0) return `${h}h ${m}m`;
-	return `${m}m`;
-}
+export type {
+	BrokerStatus,
+	MockHistoryPoint,
+	NodeInfo,
+	SessionInfo,
+	SessionSubscription,
+} from "@/lib/types";
+export {
+	generateMockHistory,
+	MOCK_NODES,
+	MOCK_SESSIONS,
+	MOCK_STATUS,
+} from "@/lib/mock/data";
+export { formatBytes, formatCount, formatUptime } from "@/lib/format";
