+++ ui/lib/format.ts
@@ -1,0 +1,21 @@
+export function formatCount(n: number): string {
+	if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
+	if (n >= 1_000) return `${(n / 1_000).toFixed(1)}K`;
+	return String(n);
+}
+
+export function formatBytes(bytes: number): string {
+	if (bytes >= 1_073_741_824) return `${(bytes / 1_073_741_824).toFixed(1)} GB`;
+	if (bytes >= 1_048_576) return `${(bytes / 1_048_576).toFixed(1)} MB`;
+	if (bytes >= 1_024) return `${(bytes / 1_024).toFixed(1)} KB`;
+	return `${bytes} B`;
+}
+
+export function formatUptime(seconds: number): string {
+	const d = Math.floor(seconds / 86400);
+	const h = Math.floor((seconds % 86400) / 3600);
+	const m = Math.floor((seconds % 3600) / 60);
+	if (d > 0) return `${d}d ${h}h`;
+	if (h > 0) return `${h}h ${m}m`;
+	return `${m}m`;
+}
